# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qtangl PQC Readiness API",
    "version": "0.9.0",
    "description": "Post-quantum readiness platform: Assess, Monitor, Convert with signed evidence."
  },
  "servers": [
    {
      "url": "https://api.qtangl.com"
    }
  ],
  "paths": {
...
    "/public/monitor-signup": {
      "post": {
        "summary": "Provision a starter monitor profile from a validated signup.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/public/lead-capture": {
      "post": {
        "summary": "Unsubscribe an email address from monitor and GTM mailing flows.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/public/onboarding-key/{token}": {
      "post": {
        "summary": "Receive Stripe webhook events for billing lifecycle synchronization.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}